requirements.txt
google-genai==1.19.0
grpcio-status==1.62.3
pillow==10.4.0
